internal/raftsnap: move "raftsnap" to internal
