+++ etcdserver/backend.go
@@ -19,11 +19,11 @@ import (
 	"os"
 	"time"
 
+	"github.com/coreos/etcd/internal/raftsnap"
 	"github.com/coreos/etcd/lease"
 	"github.com/coreos/etcd/mvcc"
 	"github.com/coreos/etcd/mvcc/backend"
 	"github.com/coreos/etcd/raft/raftpb"
-	"github.com/coreos/etcd/raftsnap"
 )
 
 func newBackend(cfg ServerConfig) backend.Backend {
