+++ etcdserver/snapshot_merge.go
@@ -17,9 +17,9 @@ package etcdserver
 import (
 	"io"
 
+	"github.com/coreos/etcd/internal/raftsnap"
 	"github.com/coreos/etcd/mvcc/backend"
 	"github.com/coreos/etcd/raft/raftpb"
-	"github.com/coreos/etcd/raftsnap"
 )
 
 // createMergedSnapshotMessage creates a snapshot message that contains: raft status (term, conf),
